breaks out morse code button logic from raspberry pi logic
clears path for making a non-GPIO adapter
